* Updated redirect URL

* Update date fields in StockPriceResponse builder

* added tracker details current price and avg

* Batch simulation with real DB persistence and skip price validation in simulation mode

- Rewrote batch-simulate endpoint to create real trackers, partitions, and
  executions in DB via DsipTrackerService and TestWorkflowService instead of
  running in-memory simulations
- Added session cookie user support with test-user fallback
- Skip execution price deviation check during simulation mode since
  day-to-day historical price swings on volatile stocks (e.g. SGMO) routinely
  exceed the 15% live-trading guard
- Removed unused executionPriceType/lockInReferenceType from BatchSimulationRequest
- Updated BatchSimulationResult with trackerId and executionLogCsvPath fields


* double value maintained in portfolio response

---------

+++ src/main/java/com/dsip/backend/dto/PortfolioResponseDto.java
@@ -18,7 +18,7 @@ public class PortfolioResponseDto {
     private Double totalMarketValue;
 
     @JsonProperty("total_capital_invested_so_far")
-    private Integer totalInvestedCapital;
+    private Double totalInvestedCapital;
 
     @JsonProperty("net_profit_percentage")
     private Double netProfitPercentage;
@@ -27,7 +27,7 @@ public class PortfolioResponseDto {
     private Double dsipTotalMarketValue;
 
     @JsonProperty("dsip_total_capital_invested_so_far")
-    private Integer dsipTotalInvestedCapital;
+    private Double dsipTotalInvestedCapital;
 
     @JsonProperty("dsip_total_net_profit_percentage")
     private Double dsipTotalNetProfitPercentage;
